services/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-docx==1.1.0
pydantic==2.5.3
python-multipart==0.0.6
langgraph>=0.0.30
langchain-core>=0.1.0

